% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decade_label}
\alias{decade_label}
\title{Label for Decade Year Range}
\usage{
decade_label(decade)
}
\arguments{
\item{decade}{start of decade as integer, e.g. 1990L}
}
\value{
character string, e.g. "1990 - 1999"
}
\description{
Label for Decade Year Range
}
