% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decade_ribbons.R
\name{decade_ribbons}
\alias{decade_ribbons}
\title{ggplot2 ribbon between q05 and q95}
\usage{
decade_ribbons(alpha = 0.1)
}
\arguments{
\item{alpha}{passed to \code{\link[ggplot2]{geom_ribbon}}}
}
\description{
ggplot2 ribbon between q05 and q95
}
