% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decades_tibble}
\alias{decades_tibble}
\title{Tibble with Decade Names and Values}
\usage{
decades_tibble(
  decade_labels,
  colors = c("darkblue", "blue", "darkgreen", "lightgreen", "orange", "red")
)
}
\arguments{
\item{decade_labels}{decade labels}

\item{values}{colour values}
}
\value{
tibble with columns \code{names} and \code{values}
}
\description{
Tibble with Decade Names and Values
}
