% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gfz_model.R
\name{download_gfz_model}
\alias{download_gfz_model}
\title{Download GFZ Model Layers}
\usage{
download_gfz_model(target_dir = tempdir())
}
\arguments{
\item{target_dir}{target directory to extract files (default: tempdir())}
}
\value{
extracts GFZ model layers in target directory
}
\description{
Download GFZ Model Layers
}
