% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{floor_decade}
\alias{floor_decade}
\title{floor to decade}
\usage{
floor_decade(value)
}
\arguments{
\item{value}{year}
}
\value{
decade
}
\description{
floors to previous decade.
}
\examples{
floor_decade(2000:2020)
}
