% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scale_x_continuous_12.R
\name{ggplot2_scale_x_continuous_12}
\alias{ggplot2_scale_x_continuous_12}
\title{Continuous x scale for months}
\usage{
ggplot2_scale_x_continuous_12()
}
\description{
Continuous x scale for months
}
