% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by_decade_month_label}
\alias{group_by_decade_month_label}
\title{Group by Dekade, Dekade_Label, Monat, Label}
\usage{
group_by_decade_month_label(data)
}
\arguments{
\item{data}{data frame}
}
\description{
Group by Dekade, Dekade_Label, Monat, Label
}
