% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numeric_quantile}
\alias{numeric_quantile}
\title{Unnamed Quantile Value}
\usage{
numeric_quantile(x, prob)
}
\arguments{
\item{x}{vector of numeric}

\item{prob}{probability passed to \code{\link{quantile}}}
}
\description{
Unnamed Quantile Value
}
