% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gfz_layer.R
\name{plot_gfz_layer}
\alias{plot_gfz_layer}
\title{Plot GFZ Model Raster Layer}
\usage{
plot_gfz_layer(layer)
}
\arguments{
\item{layer}{a layer as retrieved by \code{\link{read_gfz_layers}}}
}
\value{
plot raster layer
}
\description{
Plot GFZ Model Raster Layer
}
