% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gfz_layers.R
\name{read_gfz_layer}
\alias{read_gfz_layer}
\title{Reads a single GFZ Model Layer}
\usage{
read_gfz_layer(path, as_raster)
}
\arguments{
\item{path}{path to GFZ Model Layer}

\item{as_raster}{convert to raster (default: TRUE)}
}
\value{
imported layer file
}
\description{
Reads a single GFZ Model Layer
}
