% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_to_decade}
\alias{round_to_decade}
\title{round to decade}
\usage{
round_to_decade(values)
}
\arguments{
\item{value}{year}
}
\value{
decade
}
\description{
rounds to next decade if end year is >=5 or floors to previous
decade if end year <5.
}
\examples{
round_to_decade(2000:2020)
}
