% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_fill_decades}
\alias{scale_fill_decades}
\title{ggplot2: scale fill decades}
\usage{
scale_fill_decades(decades, ...)
}
\arguments{
\item{decades}{decades}

\item{...}{additional arguments passed to xxx}
}
\value{
re-scales ggplot2
}
\description{
ggplot2: scale fill decades
}
